feat: adjust api to v4 version

+++ app/src/main/java/com/side/project/foodmap/data/remote/api/restaurant/DistanceSearch.kt
@@ -22,7 +22,7 @@ class DistanceSearchRes(
     data class Result(
         val msg: String? = null,
         val updated: Boolean,
-        val placeCount: Long,
+        val placeCount: Int,
         val placeList: ArrayList<PlaceList>
     )
 }
