+++ app/src/main/java/com/side/project/foodmap/data/remote/api/restaurant/KeywordSearch.kt
@@ -18,7 +18,7 @@ class KeywordSearchRes(
     data class Result(
         val msg: String? = null,
         val updated: Boolean,
-        val placeCount: Long,
+        val placeCount: Int,
         val placeList: ArrayList<PlaceList>
     )
 }
